+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## WIP
 * Fix: npc wrong behavior when ball goes bottom corner.
+* Deleted: player keyboard control.
+* Added: player mouse control
 
 ## v1.0.0-alpha
 * Added: fx sounds.